Create Memory Mapping2.c
a C/C++ program that accepts a file name as a command line argument and
creates a memory map of the file. The program should then use two threads to replace all
integers in the file with spaces. The first thread will be responsible for replacing integers in the
first half of the map, while the second thread will replace integers in the second half of the map.
The file is assumed to have 100 bytes. To accomplish this, pass the map pointer to the first thread
as a parameter and add 50 to the map pointer and pass it as a parameter to the second thread.

Sample Data for File:
v1gU6OTgN7DMifG7zmQWp04ZEyGmRifq1uFsS9RzZWcCQL7jBMNKUQVEAIsKZia4
0M3TqJeGEMEkkSagfUc7mU3PbQ1zsiJm23Hq